+++ src/main/java/com/delgo/reward/repository/CertRepository.java
@@ -1,7 +1,7 @@
 package com.delgo.reward.repository;
 
 
-import com.delgo.reward.domain.Certification;
+import com.delgo.reward.domain.certification.Certification;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
 import org.springframework.data.jpa.repository.JpaRepository;
