+++ src/main/java/com/delgo/reward/service/MapService.java
@@ -1,7 +1,7 @@
 package com.delgo.reward.service;
 
 
-import com.delgo.reward.domain.Certification;
+import com.delgo.reward.domain.certification.Certification;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
